+++ plugins/plugin-github/src/components/github/repos/RepositoryIssues.tsx
@@ -76,19 +76,19 @@ const RepositoryIssues: React.FunctionComponent<IRepositoryIssuesProps> = ({
           <ToggleGroup aria-label="Type">
             <ToggleGroupItem
               className="pf-u-text-nowrap"
-              text="open"
+              text="Open"
               isSelected={state === 'open'}
               onChange={(): void => setState('open')}
             />
             <ToggleGroupItem
               className="pf-u-text-nowrap"
-              text="closed"
+              text="Closed"
               isSelected={state === 'closed'}
               onChange={(): void => setState('closed')}
             />
             <ToggleGroupItem
               className="pf-u-text-nowrap"
-              text="all"
+              text="All"
               isSelected={state === 'all'}
               onChange={(): void => setState('all')}
             />