+++ plugins/plugin-github/src/components/panel/Panel.tsx
@@ -14,6 +14,7 @@ import RepositoryWorkflowRuns from '../github/repos/RepositoryWorkflowRuns';
 import Team from '../github/teams/Team';
 import TeamMembers from '../github/teams/TeamMembers';
 import TeamRepos from '../github/teams/TeamRepos';
+import UserNotifications from '../github/users/UserNotifications';
 import UserPullRequests from '../github/users/UserPullRequests';
 
 interface IGitHubPluginPanelProps extends IPluginPanelProps {
@@ -30,7 +31,7 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 }: IGitHubPluginPanelProps) => {
   if (options && options.type && options.type === 'orgmemebers') {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <OrgMembers title={title} description={description} instance={instance} />
       </AuthContextProvider>
     );
@@ -38,7 +39,7 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 
   if (options && options.type && options.type === 'orgpullrequests') {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <OrgPullRequests title={title} description={description} instance={instance} />
       </AuthContextProvider>
     );
@@ -46,7 +47,7 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 
   if (options && options.type && options.type === 'orgrepositories') {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <OrgRepos title={title} description={description} instance={instance} setDetails={setDetails} />
       </AuthContextProvider>
     );
@@ -54,7 +55,7 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 
   if (options && options.type && options.type === 'orgteams') {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <OrgTeams title={title} description={description} instance={instance} setDetails={setDetails} />
       </AuthContextProvider>
     );
@@ -62,7 +63,7 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 
   if (options && options.type && options.type === 'team' && options.team) {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <Team title={title} description={description} slug={options.team} instance={instance} />
       </AuthContextProvider>
     );
@@ -70,7 +71,7 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 
   if (options && options.type && options.type === 'teammembers' && options.team) {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <TeamMembers title={title} description={description} slug={options.team} instance={instance} />
       </AuthContextProvider>
     );
@@ -78,7 +79,7 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 
   if (options && options.type && options.type === 'teamrepositories' && options.team) {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <TeamRepos
           title={title}
           description={description}
@@ -92,7 +93,7 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 
   if (options && options.type && options.type === 'repository' && options.repository) {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <Repository title={title} description={description} repo={options.repository} instance={instance} />
       </AuthContextProvider>
     );
@@ -100,7 +101,7 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 
   if (options && options.type && options.type === 'repositoryissues' && options.repository) {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <RepositoryIssues title={title} description={description} repo={options.repository} instance={instance} />
       </AuthContextProvider>
     );
@@ -108,7 +109,7 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 
   if (options && options.type && options.type === 'repositorypullrequests' && options.repository) {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <RepositoryPullRequests title={title} description={description} repo={options.repository} instance={instance} />
       </AuthContextProvider>
     );
@@ -116,7 +117,7 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 
   if (options && options.type && options.type === 'repositoryworkflowruns' && options.repository) {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <RepositoryWorkflowRuns
           title={title}
           description={description}
@@ -130,8 +131,16 @@ const Panel: React.FunctionComponent<IGitHubPluginPanelProps> = ({
 
   if (options && options.type && options.type === 'userpullrequests' && options.repository) {
     return (
-      <AuthContextProvider title={title} instance={instance}>
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
         <UserPullRequests title={title} description={description} instance={instance} />
+      </AuthContextProvider>
+    );
+  }
+
+  if (options && options.type && options.type === 'usernotifications' && options.repository) {
+    return (
+      <AuthContextProvider title={title} isNotification={false} instance={instance}>
+        <UserNotifications title={title} description={description} instance={instance} />
       </AuthContextProvider>
     );
   }